Joystick fully functional
The joystick is fully functional for both mouse clicks and touch events.  The two canvases have been stacked on top of each other for convenience of the user.  Necessary libs have been included so that the code can be used offline.  Some lag still exists, but it is never more than a few tens of millis.