Improved State Saving
Improved state saving by find a functional method for deep copying JavaScript objects.  It's not an eloquent method of deep copying, but it appears to be one of the best ways to do it in JS without an external library.  Interestingly this method also improved the latency issue by properly copying the incoming messages so that the original message could be garbage collected.  Other comments and documentation added.  Included a GUI text for displaying connection status so that the user doesn't have to open a console.